TipTapExperiment - 0.1 - A new plugin
I found out about TipTap.js which is a better maintained and easier to customise text editor plugin and decided to give it a try. Comapred to Draft.js it seems to much better. It's a lot easier to get started and comes with a more customisable editor. Very happy so far as im already further on that i was with draft.js in a fraction of the time