0.5.4 - Created Updating Live Text
I have created a very simple redux based system where you enter your input into the form. This is passed to the item-review area which previous the post fully styled. If the input is long enough, you can confirm the post which will update the main livetext state and update the page with a new post. It's not finished but the core functionality is in place