0.12.8 - finally a solution
it's taken me literally almost a week but I couldnt for the life of me find a solution that worked correctly. So here it is.  An overlay that appears on post creation that resets the form. It's not that elegant but It's the best I could come up with. I just want to move on with this.