0.13.4 - Smaller steps
I had a bad day trying to proceed with the next steps of this so im taking smaller steps. My terminology is also a mess so i am going to change from live texts to columns. It makes it easier to recognise the difference. A live text/ column is where posts are. Columns have different content from each other.